Things are coming along!
We can click the button and get a new random image!

This works by reading the tags on the HTML and finding the right URL and Method to use
Then we replace the specified id with the body of the request!

All this works without a page load, and no specific js for this usecase